Added Incomplete Path-Finding Sim
Added matlab simulation to demonstrate path-following algorithm.
ac_path_following takes in a vector of position vectors (x,y,theta), an
initial position vector, and the number of cycles to take.

Currently, the simulation only calculates the intersection between the
normals of the platform and the next waypoint.

TODO:
Use the intersection and angle comparison to determine the appropriate
behavior for the platform.

Determine when the current waypoint has been passed.